# Conflicts:
#	CMakeLists.txt
#	include/json_struct/json_struct.h

+++ examples/01_simple_struct.cpp
@@ -1,5 +1,5 @@
 #include <string>
-#include <json_struct.h>
+#include <json_struct/json_struct.h>
 
 const char json[] = R"json(
 {
@@ -41,7 +41,7 @@ struct TypeHandler<ModuleList>
       return JS::Error::ExpectedArrayStart;
 
     context.nextToken();
-    for (size_t i = 0; i < ModuleList::ReservedSize; i++)
+    for (int i = 0; i < int(ModuleList::ReservedSize); i++)
     {
       if (context.error != JS::Error::NoError)
         return context.error;
@@ -69,7 +69,7 @@ struct TypeHandler<ModuleList>
     serializer.write(token);
 
     token.name = DataRef("");
-    for (size_t i = 0; i < from_type.size; i++)
+    for (int i = 0; i < from_type.size; i++)
       TypeHandler<VecMember>::from(from_type.modules[i], token, serializer);
 
     token.name = DataRef("");
